backend/requirements.txt
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
yadisk==3.1.0
yarl==1.11.1
